pyproject.toml
[project]
name = "ai-news-platform"
version = "1.0.0"
description = "Local-first AI news aggregation platform"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "tenacity>=8.2.3",
    "python-dateutil>=2.8.2",
    "sqlalchemy>=2.0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
]
embeddings = [
    "sentence-transformers>=2.3.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend"]

[tool.setuptools.package-data]
backend = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
